fix: cast observations to float32 before rlax.signed_logp1 in Encoder
rlax.signed_logp1 (which replaced the hand-rolled symlog in a prior
commit on this branch) asserts its input is already float via
chex.assert_type, unlike symlog's implicit numpy-style promotion.
navix observations can be uint8 (e.g. categorical/symbolic
observations), and Dreamer.collect_experience feeds the raw env
observation straight into Encoder via WorldModel.posterior_step - this
crashed every real training step with a Chex assertion error, caught by
running tests/test_dreamer.py on CPU (test_dreamer_trains_one_update_
without_nans). Casting explicitly in Encoder restores the original
tolerance for any numeric input dtype.

+++ navix/agents/models.py
@@ -210,7 +210,11 @@ class Encoder(nn.Module):
                 nn.Dense(self.embed_size),
             ]
         )
-        return net(rlax.signed_logp1(obs))
+        # rlax.signed_logp1 (unlike the hand-rolled symlog it replaced)
+        # asserts its input is already float via chex.assert_type - navix
+        # observations can be uint8 (e.g. `categorical`/`symbolic`), so
+        # cast explicitly rather than relying on implicit promotion.
+        return net(rlax.signed_logp1(obs.astype(jnp.float32)))
 
 
 class Decoder(nn.Module):
